fix: Fix examples for Side Panle
