+++ modules/react/side-panel/stories/SidePanel.mdx
@@ -9,13 +9,14 @@ import {
 
 import SidePanelStoriesMeta from './SidePanel.stories';
 
+import {AlternativePanel} from './examples/AlternativePanel';
 import {AlwaysOpen} from './examples/AlwaysOpen';
 import {Basic} from './examples/Basic';
 import {ExternalControl} from './examples/ExternalControl';
 import {Heading} from './examples/Heading';
 import {OnStateTransition} from './examples/OnStateTransition';
+import {OverlayPanel} from './examples/OverlayPanel';
 import {RightOrigin} from './examples/RightOrigin';
-import {AlternativePanel, OverlayPanel} from './examples/Variant';
 
 <Meta of={SidePanelStoriesMeta} />
